api/requirements.txt
argon2-cffi==20.1.0
cffi==1.14.6
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonpickle==2.0.0
MarkupSafe==2.0.1
mongoengine==0.23.1
pycparser==2.20
pymongo==3.11.4
six==1.16.0
Werkzeug==2.0.1
